requirements.txt
absl-py==0.8.1
astor==0.8.0
cycler==0.10.0
Django==2.2.7
gast==0.3.2
google-pasta==0.1.8
grpcio==1.24.3
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.3
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
pyparsing==2.4.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
